added get list document api request

+++ database/migrations/2024_10_07_121710_add_column_progress_to_projects.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('projects', function (Blueprint $table) {
-            $table->integer('progress')->default(0);
+            $table->float('progress')->default(0);
         });
     }
